+++ Source/Objects/Hardware/Internet/WebRaker/ORWebRakerModel.m
@@ -224,24 +224,24 @@ NSString* ORWebRakerMaxValueChanged        = @"ORWebRakerMaxValueChanged";
                 
                 int i;
                 //make sure all the array sizes are the same as the data set
-                int lowLimitCountDiff = [data count] - [lowLimits count];
+                int lowLimitCountDiff = (int)([data count] - [lowLimits count]);
                 if(lowLimitCountDiff < 0)for(i=0;i<lowLimitCountDiff;i++)[lowLimits removeLastObject];
                 if(lowLimitCountDiff > 0)for(i=0;i<lowLimitCountDiff;i++)[lowLimits addObject:[NSNumber numberWithFloat:0]];
                 
-                int hiLimitCountDiff = [data count] - [hiLimits count];
+                int hiLimitCountDiff = (int)([data count] - [hiLimits count]);
                 if(hiLimitCountDiff < 0)for(i=0;i<hiLimitCountDiff;i++)[hiLimits removeLastObject];
                 if(hiLimitCountDiff > 0)for(i=0;i<hiLimitCountDiff;i++)[hiLimits addObject:[NSNumber numberWithFloat:0]];
          
-                int minValueCountDiff = [data count] - [minValues count];
+                int minValueCountDiff = (int)([data count] - [minValues count]);
                 if(minValueCountDiff < 0)for(i=0;i<minValueCountDiff;i++)[minValues removeLastObject];
                 if(minValueCountDiff > 0)for(i=0;i<minValueCountDiff;i++)[minValues addObject:[NSNumber numberWithFloat:0]];
 
-                int maxValueCountDiff = [data count] - [maxValues count];
+                int maxValueCountDiff = (int)([data count] - [maxValues count]);
                 if(maxValueCountDiff < 0)for(i=0;i<maxValueCountDiff;i++)[maxValues removeLastObject];
                 if(maxValueCountDiff > 0)for(i=0;i<maxValueCountDiff;i++)[maxValues addObject:[NSNumber numberWithFloat:0]];
 
                 
-                int timeRateCountDiff = [data count] - [timeRates count];
+                int timeRateCountDiff = (int)([data count] - [timeRates count]);
                 if(timeRateCountDiff < 0)for(i=0;i<timeRateCountDiff;i++)[timeRates removeLastObject];
                 if(timeRateCountDiff > 0)for(i=0;i<timeRateCountDiff;i++){
                     if(!timeRates)timeRates = [[NSMutableArray alloc] init];
